requirements.txt
absl-py==0.8.1
appdirs==1.4.3
astor==0.8.0
attrs==19.3.0
awscli==1.16.266
beautifulsoup4==4.8.1
black==19.3b0
boto3==1.10.2
botocore==1.13.2
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cssselect==1.1.0
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.3
jmespath==0.9.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.4
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
